+++ CHANGELOG.rst
@@ -15,6 +15,7 @@ Version `2.0.0`_
   * Added `similar` command to support IP Similarity lookups
   * Added `timeline` command to support IP Time Series Single Attribute lookups
   * Added `timelinehourly` command to support IP Time Series Hourly Summary lookups
+  * Added `timelinedaily` command to support IP Time Series Daily Summary lookups
 
 
 * CLI:
@@ -22,12 +23,14 @@ Version `2.0.0`_
   * Added `similar` command to support IP Similarity lookups
   * Added `timeline` command to support IP Time Series Single Attribute lookups
   * Added `timelinehourly` command to support IP Time Series Hourly Summary lookups
+  * Added `timelinedaily` command to support IP Time Series Hourly Summary lookups
 
 * Dependencies:
 
   * Updated dict2xml to version 1.7.2
   * Updated colorama to version 0.4.6
   * Update structlog to version 22.3.0
+  * Update cachetools to version 5.3.0
 
 Version `1.3.0`_
 ================
@@ -340,4 +343,4 @@ Version `0.2.0`_
 .. _`1.2.0`: https://github.com/GreyNoise-Intelligence/pygreynoise/compare/v1.1.0...1.2.0
 .. _`1.2.1`: https://github.com/GreyNoise-Intelligence/pygreynoise/compare/v1.2.0...1.2.1
 .. _`1.3.0`: https://github.com/GreyNoise-Intelligence/pygreynoise/compare/v1.2.1...1.3.0
-.. _`2.0.0`: https://github.com/GreyNoise-Intelligence/pygreynoise/compare/v1.3.0...2.0.0
+.. _`2.0.0`: https://github.com/GreyNoise-Intelligence/pygreynoise/compare/v1.3.0...HEAD